testing unconverted combination

+++ main/drivers/include/mag.h
@@ -6,7 +6,7 @@
 
 #include "i2c.h"
 
-int get_x_output(i2c_inst_t *i2c);
+int16_t get_x_output(i2c_inst_t *i2c);
 uint16_t get_x_output_raw(i2c_inst_t *i2c);
 
 int get_y_output(i2c_inst_t *i2c);
